Moved index registry into its own file
