+++ sql/index_registry_test.go
